Remove grouping of Sealed classes entirely. Not required since Kotlin 1.5
