+++ src/test/resources/examples/externalReferences/targeted/models/OneOfTwo.kt
@@ -1,0 +1,15 @@
+package examples.externalReferences.targeted.models
+
+import com.fasterxml.jackson.`annotation`.JsonProperty
+import javax.validation.constraints.NotNull
+import kotlin.String
+
+public data class OneOfTwo(
+  @param:JsonProperty("oneOfTwo")
+  @get:JsonProperty("oneOfTwo")
+  public val oneOfTwo: String? = null,
+  @get:JsonProperty("discriminator")
+  @get:NotNull
+  @param:JsonProperty("discriminator")
+  override val discriminator: String = "OneOfTwo",
+) : ParentOneOf()
