+++ src/test/resources/examples/externalReferences/targeted/models/ParentOneOf.kt
@@ -1,9 +1,7 @@
 package examples.externalReferences.targeted.models
 
-import com.fasterxml.jackson.`annotation`.JsonProperty
 import com.fasterxml.jackson.`annotation`.JsonSubTypes
 import com.fasterxml.jackson.`annotation`.JsonTypeInfo
-import javax.validation.constraints.NotNull
 import kotlin.String
 
 @JsonTypeInfo(
@@ -17,23 +15,3 @@ import kotlin.String
 public sealed class ParentOneOf() {
   public abstract val discriminator: String
 }
-
-public data class OneOfOne(
-  @param:JsonProperty("oneOfOne")
-  @get:JsonProperty("oneOfOne")
-  public val oneOfOne: String? = null,
-  @get:JsonProperty("discriminator")
-  @get:NotNull
-  @param:JsonProperty("discriminator")
-  override val discriminator: String = "OneOfOne",
-) : ParentOneOf()
-
-public data class OneOfTwo(
-  @param:JsonProperty("oneOfTwo")
-  @get:JsonProperty("oneOfTwo")
-  public val oneOfTwo: String? = null,
-  @get:JsonProperty("discriminator")
-  @get:NotNull
-  @param:JsonProperty("discriminator")
-  override val discriminator: String = "OneOfTwo",
-) : ParentOneOf()
