+++ src/test/resources/examples/oneOfPolymorphicModels/models/PolymorphicSuperTypeOne.kt
@@ -1,10 +1,7 @@
 package examples.oneOfPolymorphicModels.models
 
-import com.fasterxml.jackson.`annotation`.JsonProperty
 import com.fasterxml.jackson.`annotation`.JsonSubTypes
 import com.fasterxml.jackson.`annotation`.JsonTypeInfo
-import javax.validation.constraints.NotNull
-import kotlin.Int
 import kotlin.String
 
 @JsonTypeInfo(
@@ -19,23 +16,3 @@ import kotlin.String
 public sealed class PolymorphicSuperTypeOne() {
   public abstract val shared: String
 }
-
-public data class PolymorphicTypeOneA(
-  @param:JsonProperty("whateverA")
-  @get:JsonProperty("whateverA")
-  public val whateverA: String? = null,
-  @get:JsonProperty("shared")
-  @get:NotNull
-  @param:JsonProperty("shared")
-  override val shared: String = "PolymorphicTypeOneA",
-) : PolymorphicSuperTypeOne()
-
-public data class PolymorphicTypeOneB(
-  @param:JsonProperty("whateverB")
-  @get:JsonProperty("whateverB")
-  public val whateverB: Int? = null,
-  @get:JsonProperty("shared")
-  @get:NotNull
-  @param:JsonProperty("shared")
-  override val shared: String = "PolymorphicTypeOneB",
-) : PolymorphicSuperTypeOne()
