+++ src/test/resources/examples/oneOfPolymorphicModels/models/PolymorphicSuperTypeTwo.kt
@@ -1,10 +1,7 @@
 package examples.oneOfPolymorphicModels.models
 
-import com.fasterxml.jackson.`annotation`.JsonProperty
 import com.fasterxml.jackson.`annotation`.JsonSubTypes
 import com.fasterxml.jackson.`annotation`.JsonTypeInfo
-import javax.validation.constraints.NotNull
-import kotlin.Int
 import kotlin.String
 
 @JsonTypeInfo(
@@ -19,23 +16,3 @@ import kotlin.String
 public sealed class PolymorphicSuperTypeTwo() {
   public abstract val shared: String
 }
-
-public data class PolymorphicTypeTwoA(
-  @param:JsonProperty("whateverC")
-  @get:JsonProperty("whateverC")
-  public val whateverC: String? = null,
-  @get:JsonProperty("shared")
-  @get:NotNull
-  @param:JsonProperty("shared")
-  override val shared: String = "PolymorphicTypeTwoA",
-) : PolymorphicSuperTypeTwo()
-
-public data class PolymorphicTypeTwoB(
-  @param:JsonProperty("whateverD")
-  @get:JsonProperty("whateverD")
-  public val whateverD: Int? = null,
-  @get:JsonProperty("shared")
-  @get:NotNull
-  @param:JsonProperty("shared")
-  override val shared: String = "PolymorphicTypeTwoB",
-) : PolymorphicSuperTypeTwo()
