+++ src/test/resources/examples/polymorphicModels/models/PolymorphicTypeOneRef.kt
@@ -1,0 +1,23 @@
+package examples.polymorphicModels.models
+
+import com.fasterxml.jackson.`annotation`.JsonProperty
+import javax.validation.constraints.NotNull
+import kotlin.String
+
+public data class PolymorphicTypeOneRef(
+  @param:JsonProperty("first_name")
+  @get:JsonProperty("first_name")
+  @get:NotNull
+  override val firstName: String,
+  @param:JsonProperty("last_name")
+  @get:JsonProperty("last_name")
+  @get:NotNull
+  override val lastName: String,
+  @param:JsonProperty("child_one_name")
+  @get:JsonProperty("child_one_name")
+  public val childOneName: String? = null,
+  @get:JsonProperty("generation")
+  @get:NotNull
+  @param:JsonProperty("generation")
+  override val generation: String = "PolymorphicTypeOne",
+) : PolymorphicSuperType(firstName, lastName)
